Normalize multi-artist subtitles when searching
Roon reports playback with cosmetic `/` delimiters between each artist
for a track.  However, the search API uses commas as the delimiter.

So if the master played a track by `Thievery Corporation / Mr. Lif` the
search would fail because our results are `Thievery Corporation,  Mr. Lif`
instead.

This may not be the correct approach, but I am just stripping off
any secondary artists completely when doing the left-match search.  It
may be that we want to just replace ` / ` with `, ` instead.

We'll see.