Complete overhaul of the play/search logic
I'm really happy with these changes.  Based on today's testing I think
this will be a much more robust search path which should be able to
reliably match tracks being played by the master DJ.

The only lame part is that when we get to the final track search page,
we're just blindly grabbing the top hit.  It's possible that we might
want this to be a more clever match for songs that have a ton of
versions.  This is good for now, though.